Scripts/MoltenCore: moved Core Hound to separate script file + renamed
