+++ src/server/scripts/EasternKingdoms/BlackrockMountain/MoltenCore/boss_magmadar.cpp
@@ -19,8 +19,6 @@ EndScriptData */
 enum Texts
 {
     EMOTE_FRENZY                    = 0,
-    EMOTE_SMOLDERING                = 0,
-    EMOTE_IGNITE                    = 1,
 };
 
 enum Spells
@@ -32,8 +30,6 @@ enum Spells
     SPELL_LAVA_BOMB_EFFECT          = 20494,                    // Spawns trap GO 177704 which triggers 19428
     SPELL_LAVA_BOMB_RANGED          = 20474,                    // This calls a dummy server side effect that cast spell 20495 to spawn GO 177704 for 60s
     SPELL_LAVA_BOMB_RANGED_EFFECT   = 20495,                    // Spawns trap GO 177704 which triggers 19428
-
-    SPELL_SERRATED_BITE             = 19771,
 };
 
 enum Events
@@ -42,8 +38,6 @@ enum Events
     EVENT_PANIC,
     EVENT_LAVA_BOMB,
     EVENT_LAVA_BOMB_RANGED,
-    EVENT_SERRATED_BITE,
-    EVENT_IGNITE,
 };
 
 constexpr float MELEE_TARGET_LOOKUP_DIST = 10.0f;
@@ -144,157 +138,6 @@ public:
     }
 };
 
-// Serrated Bites timer may be wrong
-class npc_magmadar_core_hound : public CreatureScript
-{
-public:
-    npc_magmadar_core_hound() : CreatureScript("npc_magmadar_core_hound")
-    {
-    }
-
-    struct npc_magmadar_core_houndAI : public CreatureAI
-    {
-        npc_magmadar_core_houndAI(Creature* creature) : CreatureAI(creature),
-            killerGUID(0),
-            smoldering(false)
-        {
-        }
-
-        void removeFeignDeath()
-        {
-            me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PREVENT_EMOTES_FROM_CHAT_TEXT);
-            me->RemoveFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_FEIGN_DEATH);
-            me->RemoveFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_DEAD);
-            me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_DISABLE_MOVE);
-            me->RemoveUnitMovementFlag(MOVEMENTFLAG_ROOT);
-            me->ClearUnitState(UNIT_STATE_DIED);
-            me->ClearUnitState(UNIT_STATE_CANNOT_AUTOATTACK);
-            me->DisableRotate(false);
-
-            if (smoldering)
-            {
-                events.RescheduleEvent(EVENT_SERRATED_BITE, 3000);
-            }
-
-            smoldering = false;
-            killerGUID = 0;
-        }
-
-        void DamageTaken(Unit* attacker, uint32& damage, DamageEffectType /*damagetype*/, SpellSchoolMask /*damageSchoolMask*/) override
-        {
-            if (me->HealthBelowPctDamaged(0, damage))
-            {
-                if (!smoldering)
-                {
-                    killerGUID = attacker->GetGUID();
-                    events.CancelEvent(EVENT_SERRATED_BITE);
-                    events.ScheduleEvent(EVENT_IGNITE, 10000);
-                    me->SetHealth(1);
-                    me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PREVENT_EMOTES_FROM_CHAT_TEXT);
-                    me->SetFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_FEIGN_DEATH);
-                    me->SetFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_DEAD);
-                    me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_DISABLE_MOVE);
-                    me->AddUnitMovementFlag(MOVEMENTFLAG_ROOT);
-                    me->AddUnitState(UNIT_STATE_DIED);
-                    me->AddUnitState(UNIT_STATE_CANNOT_AUTOATTACK);
-                    me->DisableRotate(true);
-                    Talk(EMOTE_SMOLDERING);
-                }
-                damage = 0;
-                smoldering = true;
-            }
-        }
-
-        void Reset() override
-        {
-            removeFeignDeath();
-        }
-
-        void JustDied(Unit* /*killer*/) override
-        {
-            removeFeignDeath();
-        }
-
-        void EnterCombat(Unit* /*victim*/) override
-        {
-            events.ScheduleEvent(EVENT_SERRATED_BITE, 3000);
-        }
-
-        void UpdateAI(uint32 diff) override
-        {
-            if (!UpdateVictim() && !smoldering)
-            {
-                return;
-            }
-
-            events.Update(diff);
-
-            while (uint32 const eventId = events.ExecuteEvent())
-            {
-                switch (eventId)
-                {
-                    case EVENT_SERRATED_BITE:
-                    {
-                        if (!smoldering)
-                        {
-                            DoCastVictim(SPELL_SERRATED_BITE);
-                        }
-                        events.RepeatEvent(urand(5000, 6000));
-                        break;
-                    }
-                    case EVENT_IGNITE:
-                    {
-                        smoldering = false;
-                        std::list<Creature*> hounds;
-                        me->GetCreaturesWithEntryInRange(hounds, 80, NPC_CORE_HOUND);
-                        hounds.remove_if([](Creature* hound) -> bool
-                        {
-                            return !hound || hound->isDead() || !hound->IsInCombat() || hound->HasUnitState(UNIT_STATE_DIED);
-                        });
-
-                        if (!hounds.empty())
-                        {
-                            Talk(EMOTE_IGNITE);
-                            me->SetFullHealth();
-                            me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PREVENT_EMOTES_FROM_CHAT_TEXT);
-                            me->RemoveFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_FEIGN_DEATH);
-                            me->RemoveFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_DEAD);
-                            me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_DISABLE_MOVE);
-                            me->RemoveUnitMovementFlag(MOVEMENTFLAG_ROOT);
-                            me->ClearUnitState(UNIT_STATE_DIED);
-                            me->ClearUnitState(UNIT_STATE_CANNOT_AUTOATTACK);
-                            me->DisableRotate(false);
-                            if (Unit* victim = SelectTarget(SELECT_TARGET_RANDOM, 0, 0.0f, true))
-                            {
-                                AttackStart(victim);
-                            }
-                        }
-                        else if (me->HasUnitState(UNIT_STATE_DIED))
-                        {
-                            Unit* killer = ObjectAccessor::GetUnit(*me, killerGUID);
-                            me->Kill(killer ? killer : me, me);
-                            killerGUID = 0;
-                        }
-                        break;
-                    }
-                }
-            }
-
-            DoMeleeAttackIfReady();
-        }
-
-    private:
-        EventMap events;
-        uint64 killerGUID;
-        bool smoldering;
-    };
-
-    CreatureAI* GetAI(Creature* creature) const override
-    {
-        return GetMoltenCoreAI<npc_magmadar_core_houndAI>(creature);
-    }
-};
-
 // 19411 Lava Bomb
 // 20474 Lava Bomb
 class spell_magmadar_lava_bomb : public SpellScriptLoader
@@ -349,6 +192,5 @@ public:
 void AddSC_boss_magmadar()
 {
     new boss_magmadar();
-    new npc_magmadar_core_hound();
     new spell_magmadar_lava_bomb();
 }
